refactor(#487): remove AS features
