+++ annotation-interface/.env.example
@@ -4,5 +4,3 @@ DB_HOST=localhost
 DB_PORT=27017
 DB_USER=changeme
 DB_PASSWORD=changeme
-
-AS_API=localhost:3000/api/v1
