+++ annotation-interface/database/models/TextBrick.ts
@@ -1,18 +1,12 @@
-import mongoose, { FilterQuery, Types } from 'mongoose';
+import mongoose, { Types } from 'mongoose';
 
 import {
     brickUsages,
     SupportedLanguage,
     BrickUsage,
     supportedLanguages,
-    pharMeLanguage,
 } from '../../common/definitions';
 import { translationIsValid } from '../helpers/brick-translations';
-import {
-    BrickResolver,
-    resolveBricks,
-    ResolvedBrick,
-} from '../helpers/resolve-bricks';
 import { IBaseModel, OptionalId } from '../helpers/types';
 
 export interface ITextBrickTranslation<IdT extends OptionalId = undefined>
@@ -29,13 +23,7 @@ export interface ITextBrick<IdT extends OptionalId = undefined>
 export type ITextBrick_DB = ITextBrick<Types.ObjectId>;
 export type ITextBrick_Str = ITextBrick<string>;
 
-export interface TextBrickModel extends mongoose.Model<ITextBrick_DB> {
-    findResolved(
-        resolver: BrickResolver,
-        filter: FilterQuery<ITextBrick_DB>,
-        language?: SupportedLanguage,
-    ): Promise<ResolvedBrick<Types.ObjectId>[]>;
-}
+export type TextBrickModel = mongoose.Model<ITextBrick_DB>;
 
 const textBrickSchema = new mongoose.Schema<ITextBrick_DB>({
     usage: {
@@ -84,18 +72,6 @@ const textBrickSchema = new mongoose.Schema<ITextBrick_DB>({
         ],
     },
 });
-
-textBrickSchema.static(
-    'findResolved',
-    async function (
-        resolver: BrickResolver,
-        filter: FilterQuery<ITextBrick_DB>,
-        language?: SupportedLanguage,
-    ): Promise<ResolvedBrick<Types.ObjectId>[]> {
-        const bricks = await this.find(filter).lean().exec();
-        return resolveBricks(resolver, bricks, language ?? pharMeLanguage);
-    },
-);
 
 // prevent client side from trying to use node module
 export default !mongoose.models
